rename result_set files to results
