+++ google/cloud/spanner/results.h
